Correccion de bugs a lo largo de la libreria, despues de realizar una prueba a detalle por primera vez de la funcionalidad de la misma usando el motor de base de datos MSSQL. Aunado a lo anterior, se hicieron mas pruebas con la serializacion a excel y archivos dando resultados positivos, pero tambien se corrigieron algunos bugs.

+++ Tools/SimpleConverter.cs
@@ -51,7 +51,7 @@ namespace DataManagement.Tools
                         }
                         catch (Exception)
                         {
-                            return DateTime.FromOADate(long.Parse(value));
+                            return DateTime.FromOADate(double.Parse(value));
                         }
                     default:
                         return Convert.ChangeType(value, underlyingType);
